+++ src/fides/api/tasks/storage.py
@@ -5,7 +5,7 @@ import os
 import secrets
 import zipfile
 from io import BytesIO
-from typing import Any, Dict, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional, Union
 
 import pandas as pd
 from botocore.exceptions import ClientError, ParamValidationError
@@ -13,11 +13,15 @@ from fideslang.validation import AnyHttpUrlString
 from loguru import logger
 
 from fides.api.cryptography.cryptographic_util import bytes_to_b64_str
-from fides.api.models.privacy_request import PrivacyRequest
 from fides.api.schemas.storage.storage import ResponseFormat, StorageSecrets
 from fides.api.service.privacy_request.dsr_package.dsr_report_builder import (
     DsrReportBuilder,
 )
+from fides.api.service.storage.s3 import (
+    create_presigned_url_for_s3,
+    generic_upload_to_s3,
+)
+from fides.api.service.storage.util import LOCAL_FIDES_UPLOAD_DIRECTORY
 from fides.api.util.aws_util import get_s3_client
 from fides.api.util.cache import get_cache, get_encryption_cache_key
 from fides.api.util.encryption.aes_gcm_encryption_scheme import (
@@ -26,7 +30,8 @@ from fides.api.util.encryption.aes_gcm_encryption_scheme import (
 from fides.api.util.storage_util import storage_json_encoder
 from fides.config import CONFIG
 
-LOCAL_FIDES_UPLOAD_DIRECTORY = "fides_uploads"
+if TYPE_CHECKING:
+    from fides.api.models.privacy_request import PrivacyRequest
 
 
 def encrypt_access_request_results(data: Union[str, bytes], request_id: str) -> str:
@@ -101,105 +106,6 @@ def write_to_in_memory_buffer(
     raise NotImplementedError(f"No handling for response format {resp_format}.")
 
 
-def create_presigned_url_for_s3(
-    s3_client: Any, bucket_name: str, file_key: str
-) -> AnyHttpUrlString:
-    """ "Generate a presigned URL to share an S3 object
-
-    :param s3_client: s3 base client
-    :param bucket_name: string
-    :param file_key: string
-    :return: Presigned URL as string.
-    """
-    response = s3_client.generate_presigned_url(
-        "get_object",
-        Params={"Bucket": bucket_name, "Key": file_key},
-        ExpiresIn=CONFIG.security.subject_request_download_link_ttl_seconds,
-    )
-
-    # The response contains the presigned URL
-    return response
-
-
-def generic_upload_to_s3(  # pylint: disable=R0913
-    storage_secrets: Dict[StorageSecrets, Any],
-    bucket_name: str,
-    file_key: str,
-    auth_method: str,
-    document: bytes,
-) -> Optional[AnyHttpUrlString]:
-    """Uploads arbitrary data to s3 returned from an access request"""
-    logger.info("Starting S3 Upload of {}", file_key)
-
-    try:
-        s3_client = get_s3_client(auth_method, storage_secrets)
-        try:
-            s3_client.put_object(Bucket=bucket_name, Key=file_key, Body=document)
-        except Exception as e:
-            logger.error("Encountered error while uploading s3 object: {}", e)
-            raise e
-
-        presigned_url: AnyHttpUrlString = create_presigned_url_for_s3(
-            s3_client, bucket_name, file_key
-        )
-
-        return presigned_url
-    except ClientError as e:
-        logger.error(
-            "Encountered error while uploading and generating link for s3 object: {}", e
-        )
-        raise e
-    except ParamValidationError as e:
-        raise ValueError(f"The parameters you provided are incorrect: {e}")
-
-
-def generic_retrieve_from_s3(
-    storage_secrets: Dict[StorageSecrets, Any],
-    bucket_name: str,
-    file_key: str,
-    auth_method: str,
-) -> Optional[bytes]:
-    """Retrieves arbitrary data from s3"""
-    logger.info("Starting S3 Retrieve of {}", file_key)
-
-    try:
-        s3_client = get_s3_client(auth_method, storage_secrets)
-        try:
-            response = s3_client.get_object(Bucket=bucket_name, Key=file_key)
-            return response["Body"].read()
-        except Exception as e:
-            logger.error("Encountered error while retrieving s3 object: {}", e)
-            raise e
-    except ClientError as e:
-        logger.error("Encountered error while retrieving s3 object: {}", e)
-        raise e
-    except ParamValidationError as e:
-        raise ValueError(f"The parameters you provided are incorrect: {e}")
-
-
-def generic_delete_from_s3(
-    storage_secrets: Dict[StorageSecrets, Any],
-    bucket_name: str,
-    file_key: str,
-    auth_method: str,
-) -> None:
-    """Deletes arbitrary data from s3"""
-    logger.info("Starting S3 Delete of {}", file_key)
-
-    try:
-        s3_client = get_s3_client(auth_method, storage_secrets)
-        try:
-            s3_client.delete_object(Bucket=bucket_name, Key=file_key)
-        except Exception as e:
-            logger.error("Encountered error while deleting s3 object: {}", e)
-            raise e
-    except ClientError as e:
-        logger.error("Encountered error while deleting s3 object: {}", e)
-        raise e
-    except ParamValidationError as e:
-        raise ValueError(f"The parameters you provided are incorrect: {e}")
-
-
 def upload_to_s3(  # pylint: disable=R0913
     storage_secrets: Dict[StorageSecrets, Any],
     data: Dict,
@@ -247,13 +153,6 @@ def upload_to_s3(  # pylint: disable=R0913
         raise e
     except ParamValidationError as e:
         raise ValueError(f"The parameters you provided are incorrect: {e}")
-
-
-def get_local_filename(file_key: str) -> str:
-    """Verifies that the local storage directory exists"""
-    if not os.path.exists(LOCAL_FIDES_UPLOAD_DIRECTORY):
-        os.makedirs(LOCAL_FIDES_UPLOAD_DIRECTORY)
-    return f"{LOCAL_FIDES_UPLOAD_DIRECTORY}/{file_key}"
 
 
 def upload_to_local(
