+++ tests/ctl/models/conftest.py
@@ -1,6 +1,4 @@
-import boto3
 import pytest
-from moto import mock_aws
 
 from fides.api.models.attachment import (
     Attachment,
@@ -9,19 +7,6 @@ from fides.api.models.attachment import (
     AttachmentType,
 )
 from fides.api.schemas.storage.storage import StorageDetails
-
-
-@pytest.fixture
-def s3_client(storage_config):
-    with mock_aws():
-        session = boto3.Session(
-            aws_access_key_id="fake_access_key",
-            aws_secret_access_key="fake_secret_key",
-            region_name="us-east-1",
-        )
-        s3 = session.client("s3")
-        s3.create_bucket(Bucket=storage_config.details[StorageDetails.BUCKET.value])
-        yield s3
 
 
 @pytest.fixture
@@ -42,7 +27,7 @@ def attachment(s3_client, db, attachment_data, monkeypatch):
     def mock_get_s3_client(auth_method, storage_secrets):
         return s3_client
 
-    monkeypatch.setattr("fides.api.tasks.storage.get_s3_client", mock_get_s3_client)
+    monkeypatch.setattr("fides.api.service.storage.s3.get_s3_client", mock_get_s3_client)
     attachment = Attachment.create_and_upload(
         db, data=attachment_data, attachment_file=b"file content"
     )
@@ -57,7 +42,7 @@ def multiple_attachments(s3_client, db, attachment_data, user, monkeypatch):
     def mock_get_s3_client(auth_method, storage_secrets):
         return s3_client
 
-    monkeypatch.setattr("fides.api.tasks.storage.get_s3_client", mock_get_s3_client)
+    monkeypatch.setattr("fides.api.service.storage.s3.get_s3_client", mock_get_s3_client)
 
     attachment_data["user_id"] = user.id
     attachment_data["file_name"] = "file_1.txt"
